Move routes to App
