+++ packages/frontend/src/App/LoggedOutApp.tsx
@@ -1,8 +1,8 @@
 import { lazy } from 'react';
 import { Route, Switch } from 'react-router-dom';
-import Header from '../Header';
+import Header from '../components/Header';
 
-const LoginPage = lazy(() => import('../../pages/LoginPage'));
+const LoginPage = lazy(() => import('../pages/LoginPage'));
 
 export default function LoggedOutApp() {
   return (
