--- packages/frontend/src/components/BodyArea/__tests__/index.test.tsx
@@ -1,13 +1,0 @@
-import BodyArea from '..';
-import { render, screen } from '../../../util/test-utils';
-
-describe('BodyArea', () => {
-  it('should render', () => {
-    render(
-      <BodyArea>
-        <span data-testid="example">Example</span>
-      </BodyArea>,
-    );
-    expect(screen.getByTestId('example')).toHaveTextContent('Example');
-  });
-});
