--- packages/frontend/src/components/BodyArea/index.css
@@ -1,4 +1,0 @@
-.BodyArea {
-  max-width: 70em;
-  margin: 0 auto;
-}
